Rename ckb-combine-lock-v2 to ckb-combine-lock
