+++ ckb-combine-lock-common/combine_lock.mol
@@ -1,17 +1,26 @@
 import blockchain;
 
-array Uint16 [byte; 2];
-
 table ChildScript {
     code_hash:      Byte32,
     hash_type:      byte,
     args:           Bytes,
 }
 
-vector ChildScriptVec <ChildScript>;
+vector ChildScriptVec <byte>;
+
+vector ChildScriptVecVec <ChildScriptVec>;
+
+vector ChildScriptArray <ChildScript>;
+
+table ChildScriptConfig {
+    array: ChildScriptArray,
+    index: ChildScriptVecVec,
+}
+
+option ChildScriptConfigOpt (ChildScriptConfig);
 
 table CombineLockWitness {
-    witness_base_index: Uint16,
-    proof: Bytes,
-    scripts: ChildScriptVec,
+    index: byte,
+    inner_witness: BytesVec,
+    script_config: ChildScriptConfigOpt,
 }
