+++ ckb-combine-lock-common/src/combine_lock_mol.rs
@@ -1,182 +1,7 @@
-// Generated by Molecule 0.7.3
+// Generated by Molecule 0.7.5
 
 use super::blockchain::*;
 use molecule::prelude::*;
-#[derive(Clone)]
-pub struct Uint16(molecule::bytes::Bytes);
-impl ::core::fmt::LowerHex for Uint16 {
-    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
-        use molecule::hex_string;
-        if f.alternate() {
-            write!(f, "0x")?;
-        }
-        write!(f, "{}", hex_string(self.as_slice()))
-    }
-}
-impl ::core::fmt::Debug for Uint16 {
-    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
-        write!(f, "{}({:#x})", Self::NAME, self)
-    }
-}
-impl ::core::fmt::Display for Uint16 {
-    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
-        use molecule::hex_string;
-        let raw_data = hex_string(&self.raw_data());
-        write!(f, "{}(0x{})", Self::NAME, raw_data)
-    }
-}
-impl ::core::default::Default for Uint16 {
-    fn default() -> Self {
-        let v: Vec<u8> = vec![0, 0];
-        Uint16::new_unchecked(v.into())
-    }
-}
-impl Uint16 {
-    pub const TOTAL_SIZE: usize = 2;
-    pub const ITEM_SIZE: usize = 1;
-    pub const ITEM_COUNT: usize = 2;
-    pub fn nth0(&self) -> Byte {
-        Byte::new_unchecked(self.0.slice(0..1))
-    }
-    pub fn nth1(&self) -> Byte {
-        Byte::new_unchecked(self.0.slice(1..2))
-    }
-    pub fn raw_data(&self) -> molecule::bytes::Bytes {
-        self.as_bytes()
-    }
-    pub fn as_reader<'r>(&'r self) -> Uint16Reader<'r> {
-        Uint16Reader::new_unchecked(self.as_slice())
-    }
-}
-impl molecule::prelude::Entity for Uint16 {
-    type Builder = Uint16Builder;
-    const NAME: &'static str = "Uint16";
-    fn new_unchecked(data: molecule::bytes::Bytes) -> Self {
-        Uint16(data)
-    }
-    fn as_bytes(&self) -> molecule::bytes::Bytes {
-        self.0.clone()
-    }
-    fn as_slice(&self) -> &[u8] {
-        &self.0[..]
-    }
-    fn from_slice(slice: &[u8]) -> molecule::error::VerificationResult<Self> {
-        Uint16Reader::from_slice(slice).map(|reader| reader.to_entity())
-    }
-    fn from_compatible_slice(slice: &[u8]) -> molecule::error::VerificationResult<Self> {
-        Uint16Reader::from_compatible_slice(slice).map(|reader| reader.to_entity())
-    }
-    fn new_builder() -> Self::Builder {
-        ::core::default::Default::default()
-    }
-    fn as_builder(self) -> Self::Builder {
-        Self::new_builder().set([self.nth0(), self.nth1()])
-    }
-}
-#[derive(Clone, Copy)]
-pub struct Uint16Reader<'r>(&'r [u8]);
-impl<'r> ::core::fmt::LowerHex for Uint16Reader<'r> {
-    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
-        use molecule::hex_string;
-        if f.alternate() {
-            write!(f, "0x")?;
-        }
-        write!(f, "{}", hex_string(self.as_slice()))
-    }
-}
-impl<'r> ::core::fmt::Debug for Uint16Reader<'r> {
-    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
-        write!(f, "{}({:#x})", Self::NAME, self)
-    }
-}
-impl<'r> ::core::fmt::Display for Uint16Reader<'r> {
-    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
-        use molecule::hex_string;
-        let raw_data = hex_string(&self.raw_data());
-        write!(f, "{}(0x{})", Self::NAME, raw_data)
-    }
-}
-impl<'r> Uint16Reader<'r> {
-    pub const TOTAL_SIZE: usize = 2;
-    pub const ITEM_SIZE: usize = 1;
-    pub const ITEM_COUNT: usize = 2;
-    pub fn nth0(&self) -> ByteReader<'r> {
-        ByteReader::new_unchecked(&self.as_slice()[0..1])
-    }
-    pub fn nth1(&self) -> ByteReader<'r> {
-        ByteReader::new_unchecked(&self.as_slice()[1..2])
-    }
-    pub fn raw_data(&self) -> &'r [u8] {
-        self.as_slice()
-    }
-}
-impl<'r> molecule::prelude::Reader<'r> for Uint16Reader<'r> {
-    type Entity = Uint16;
-    const NAME: &'static str = "Uint16Reader";
-    fn to_entity(&self) -> Self::Entity {
-        Self::Entity::new_unchecked(self.as_slice().to_owned().into())
-    }
-    fn new_unchecked(slice: &'r [u8]) -> Self {
-        Uint16Reader(slice)
-    }
-    fn as_slice(&self) -> &'r [u8] {
-        self.0
-    }
-    fn verify(slice: &[u8], _compatible: bool) -> molecule::error::VerificationResult<()> {
-        use molecule::verification_error as ve;
-        let slice_len = slice.len();
-        if slice_len != Self::TOTAL_SIZE {
-            return ve!(Self, TotalSizeNotMatch, Self::TOTAL_SIZE, slice_len);
-        }
-        Ok(())
-    }
-}
-pub struct Uint16Builder(pub(crate) [Byte; 2]);
-impl ::core::fmt::Debug for Uint16Builder {
-    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
-        write!(f, "{}({:?})", Self::NAME, &self.0[..])
-    }
-}
-impl ::core::default::Default for Uint16Builder {
-    fn default() -> Self {
-        Uint16Builder([Byte::default(), Byte::default()])
-    }
-}
-impl Uint16Builder {
-    pub const TOTAL_SIZE: usize = 2;
-    pub const ITEM_SIZE: usize = 1;
-    pub const ITEM_COUNT: usize = 2;
-    pub fn set(mut self, v: [Byte; 2]) -> Self {
-        self.0 = v;
-        self
-    }
-    pub fn nth0(mut self, v: Byte) -> Self {
-        self.0[0] = v;
-        self
-    }
-    pub fn nth1(mut self, v: Byte) -> Self {
-        self.0[1] = v;
-        self
-    }
-}
-impl molecule::prelude::Builder for Uint16Builder {
-    type Entity = Uint16;
-    const NAME: &'static str = "Uint16Builder";
-    fn expected_length(&self) -> usize {
-        Self::TOTAL_SIZE
-    }
-    fn write<W: molecule::io::Write>(&self, writer: &mut W) -> molecule::io::Result<()> {
-        writer.write_all(self.0[0].as_slice())?;
-        writer.write_all(self.0[1].as_slice())?;
-        Ok(())
-    }
-    fn build(&self) -> Self::Entity {
-        let mut inner = Vec::with_capacity(self.expected_length());
-        self.write(&mut inner)
-            .unwrap_or_else(|_| panic!("{} build should be ok", Self::NAME));
-        Uint16::new_unchecked(inner.into())
-    }
-}
 #[derive(Clone)]
 pub struct ChildScript(molecule::bytes::Bytes);
 impl ::core::fmt::LowerHex for ChildScript {
@@ -208,14 +33,15 @@ impl ::core::fmt::Display for ChildScript {
 }
 impl ::core::default::Default for ChildScript {
     fn default() -> Self {
-        let v: Vec<u8> = vec![
-            53, 0, 0, 0, 16, 0, 0, 0, 48, 0, 0, 0, 49, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-            0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-        ];
-        ChildScript::new_unchecked(v.into())
+        let v = molecule::bytes::Bytes::from_static(&Self::DEFAULT_VALUE);
+        ChildScript::new_unchecked(v)
     }
 }
 impl ChildScript {
+    const DEFAULT_VALUE: [u8; 53] = [
+        53, 0, 0, 0, 16, 0, 0, 0, 48, 0, 0, 0, 49, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+    ];
     pub const FIELD_COUNT: usize = 3;
     pub fn total_size(&self) -> usize {
         molecule::unpack_number(self.as_slice()) as usize
@@ -379,9 +205,6 @@ impl<'r> molecule::prelude::Reader<'r> for ChildScriptReader<'r> {
         if slice_len != total_size {
             return ve!(Self, TotalSizeNotMatch, total_size, slice_len);
         }
-        if slice_len == molecule::NUMBER_SIZE && Self::FIELD_COUNT == 0 {
-            return Ok(());
-        }
         if slice_len < molecule::NUMBER_SIZE * 2 {
             return ve!(Self, HeaderIsBroken, molecule::NUMBER_SIZE * 2, slice_len);
         }
@@ -485,33 +308,25 @@ impl ::core::fmt::Debug for ChildScriptVec {
 }
 impl ::core::fmt::Display for ChildScriptVec {
     fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
-        write!(f, "{} [", Self::NAME)?;
-        for i in 0..self.len() {
-            if i == 0 {
-                write!(f, "{}", self.get_unchecked(i))?;
-            } else {
-                write!(f, ", {}", self.get_unchecked(i))?;
-            }
-        }
-        write!(f, "]")
+        use molecule::hex_string;
+        let raw_data = hex_string(&self.raw_data());
+        write!(f, "{}(0x{})", Self::NAME, raw_data)
     }
 }
 impl ::core::default::Default for ChildScriptVec {
     fn default() -> Self {
-        let v: Vec<u8> = vec![4, 0, 0, 0];
-        ChildScriptVec::new_unchecked(v.into())
+        let v = molecule::bytes::Bytes::from_static(&Self::DEFAULT_VALUE);
+        ChildScriptVec::new_unchecked(v)
     }
 }
 impl ChildScriptVec {
+    const DEFAULT_VALUE: [u8; 4] = [0, 0, 0, 0];
+    pub const ITEM_SIZE: usize = 1;
     pub fn total_size(&self) -> usize {
-        molecule::unpack_number(self.as_slice()) as usize
+        molecule::NUMBER_SIZE + Self::ITEM_SIZE * self.item_count()
     }
     pub fn item_count(&self) -> usize {
-        if self.total_size() == molecule::NUMBER_SIZE {
-            0
-        } else {
-            (molecule::unpack_number(&self.as_slice()[molecule::NUMBER_SIZE..]) as usize / 4) - 1
-        }
+        molecule::unpack_number(self.as_slice()) as usize
     }
     pub fn len(&self) -> usize {
         self.item_count()
@@ -519,24 +334,20 @@ impl ChildScriptVec {
     pub fn is_empty(&self) -> bool {
         self.len() == 0
     }
-    pub fn get(&self, idx: usize) -> Option<ChildScript> {
+    pub fn get(&self, idx: usize) -> Option<Byte> {
         if idx >= self.len() {
             None
         } else {
             Some(self.get_unchecked(idx))
         }
     }
-    pub fn get_unchecked(&self, idx: usize) -> ChildScript {
-        let slice = self.as_slice();
-        let start_idx = molecule::NUMBER_SIZE * (1 + idx);
-        let start = molecule::unpack_number(&slice[start_idx..]) as usize;
-        if idx == self.len() - 1 {
-            ChildScript::new_unchecked(self.0.slice(start..))
-        } else {
-            let end_idx = start_idx + molecule::NUMBER_SIZE;
-            let end = molecule::unpack_number(&slice[end_idx..]) as usize;
-            ChildScript::new_unchecked(self.0.slice(start..end))
-        }
+    pub fn get_unchecked(&self, idx: usize) -> Byte {
+        let start = molecule::NUMBER_SIZE + Self::ITEM_SIZE * idx;
+        let end = start + Self::ITEM_SIZE;
+        Byte::new_unchecked(self.0.slice(start..end))
+    }
+    pub fn raw_data(&self) -> molecule::bytes::Bytes {
+        self.0.slice(molecule::NUMBER_SIZE..)
     }
     pub fn as_reader<'r>(&'r self) -> ChildScriptVecReader<'r> {
         ChildScriptVecReader::new_unchecked(self.as_slice())
@@ -585,6 +396,161 @@ impl<'r> ::core::fmt::Debug for ChildScriptVecReader<'r> {
 }
 impl<'r> ::core::fmt::Display for ChildScriptVecReader<'r> {
     fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        use molecule::hex_string;
+        let raw_data = hex_string(&self.raw_data());
+        write!(f, "{}(0x{})", Self::NAME, raw_data)
+    }
+}
+impl<'r> ChildScriptVecReader<'r> {
+    pub const ITEM_SIZE: usize = 1;
+    pub fn total_size(&self) -> usize {
+        molecule::NUMBER_SIZE + Self::ITEM_SIZE * self.item_count()
+    }
+    pub fn item_count(&self) -> usize {
+        molecule::unpack_number(self.as_slice()) as usize
+    }
+    pub fn len(&self) -> usize {
+        self.item_count()
+    }
+    pub fn is_empty(&self) -> bool {
+        self.len() == 0
+    }
+    pub fn get(&self, idx: usize) -> Option<ByteReader<'r>> {
+        if idx >= self.len() {
+            None
+        } else {
+            Some(self.get_unchecked(idx))
+        }
+    }
+    pub fn get_unchecked(&self, idx: usize) -> ByteReader<'r> {
+        let start = molecule::NUMBER_SIZE + Self::ITEM_SIZE * idx;
+        let end = start + Self::ITEM_SIZE;
+        ByteReader::new_unchecked(&self.as_slice()[start..end])
+    }
+    pub fn raw_data(&self) -> &'r [u8] {
+        &self.as_slice()[molecule::NUMBER_SIZE..]
+    }
+}
+impl<'r> molecule::prelude::Reader<'r> for ChildScriptVecReader<'r> {
+    type Entity = ChildScriptVec;
+    const NAME: &'static str = "ChildScriptVecReader";
+    fn to_entity(&self) -> Self::Entity {
+        Self::Entity::new_unchecked(self.as_slice().to_owned().into())
+    }
+    fn new_unchecked(slice: &'r [u8]) -> Self {
+        ChildScriptVecReader(slice)
+    }
+    fn as_slice(&self) -> &'r [u8] {
+        self.0
+    }
+    fn verify(slice: &[u8], _compatible: bool) -> molecule::error::VerificationResult<()> {
+        use molecule::verification_error as ve;
+        let slice_len = slice.len();
+        if slice_len < molecule::NUMBER_SIZE {
+            return ve!(Self, HeaderIsBroken, molecule::NUMBER_SIZE, slice_len);
+        }
+        let item_count = molecule::unpack_number(slice) as usize;
+        if item_count == 0 {
+            if slice_len != molecule::NUMBER_SIZE {
+                return ve!(Self, TotalSizeNotMatch, molecule::NUMBER_SIZE, slice_len);
+            }
+            return Ok(());
+        }
+        let total_size = molecule::NUMBER_SIZE + Self::ITEM_SIZE * item_count;
+        if slice_len != total_size {
+            return ve!(Self, TotalSizeNotMatch, total_size, slice_len);
+        }
+        Ok(())
+    }
+}
+#[derive(Debug, Default)]
+pub struct ChildScriptVecBuilder(pub(crate) Vec<Byte>);
+impl ChildScriptVecBuilder {
+    pub const ITEM_SIZE: usize = 1;
+    pub fn set(mut self, v: Vec<Byte>) -> Self {
+        self.0 = v;
+        self
+    }
+    pub fn push(mut self, v: Byte) -> Self {
+        self.0.push(v);
+        self
+    }
+    pub fn extend<T: ::core::iter::IntoIterator<Item = Byte>>(mut self, iter: T) -> Self {
+        for elem in iter {
+            self.0.push(elem);
+        }
+        self
+    }
+    pub fn replace(&mut self, index: usize, v: Byte) -> Option<Byte> {
+        self.0
+            .get_mut(index)
+            .map(|item| ::core::mem::replace(item, v))
+    }
+}
+impl molecule::prelude::Builder for ChildScriptVecBuilder {
+    type Entity = ChildScriptVec;
+    const NAME: &'static str = "ChildScriptVecBuilder";
+    fn expected_length(&self) -> usize {
+        molecule::NUMBER_SIZE + Self::ITEM_SIZE * self.0.len()
+    }
+    fn write<W: molecule::io::Write>(&self, writer: &mut W) -> molecule::io::Result<()> {
+        writer.write_all(&molecule::pack_number(self.0.len() as molecule::Number))?;
+        for inner in &self.0[..] {
+            writer.write_all(inner.as_slice())?;
+        }
+        Ok(())
+    }
+    fn build(&self) -> Self::Entity {
+        let mut inner = Vec::with_capacity(self.expected_length());
+        self.write(&mut inner)
+            .unwrap_or_else(|_| panic!("{} build should be ok", Self::NAME));
+        ChildScriptVec::new_unchecked(inner.into())
+    }
+}
+pub struct ChildScriptVecIterator(ChildScriptVec, usize, usize);
+impl ::core::iter::Iterator for ChildScriptVecIterator {
+    type Item = Byte;
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.1 >= self.2 {
+            None
+        } else {
+            let ret = self.0.get_unchecked(self.1);
+            self.1 += 1;
+            Some(ret)
+        }
+    }
+}
+impl ::core::iter::ExactSizeIterator for ChildScriptVecIterator {
+    fn len(&self) -> usize {
+        self.2 - self.1
+    }
+}
+impl ::core::iter::IntoIterator for ChildScriptVec {
+    type Item = Byte;
+    type IntoIter = ChildScriptVecIterator;
+    fn into_iter(self) -> Self::IntoIter {
+        let len = self.len();
+        ChildScriptVecIterator(self, 0, len)
+    }
+}
+#[derive(Clone)]
+pub struct ChildScriptVecVec(molecule::bytes::Bytes);
+impl ::core::fmt::LowerHex for ChildScriptVecVec {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        use molecule::hex_string;
+        if f.alternate() {
+            write!(f, "0x")?;
+        }
+        write!(f, "{}", hex_string(self.as_slice()))
+    }
+}
+impl ::core::fmt::Debug for ChildScriptVecVec {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        write!(f, "{}({:#x})", Self::NAME, self)
+    }
+}
+impl ::core::fmt::Display for ChildScriptVecVec {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
         write!(f, "{} [", Self::NAME)?;
         for i in 0..self.len() {
             if i == 0 {
@@ -596,7 +562,448 @@ impl<'r> ::core::fmt::Display for ChildScriptVecReader<'r> {
         write!(f, "]")
     }
 }
-impl<'r> ChildScriptVecReader<'r> {
+impl ::core::default::Default for ChildScriptVecVec {
+    fn default() -> Self {
+        let v = molecule::bytes::Bytes::from_static(&Self::DEFAULT_VALUE);
+        ChildScriptVecVec::new_unchecked(v)
+    }
+}
+impl ChildScriptVecVec {
+    const DEFAULT_VALUE: [u8; 4] = [4, 0, 0, 0];
+    pub fn total_size(&self) -> usize {
+        molecule::unpack_number(self.as_slice()) as usize
+    }
+    pub fn item_count(&self) -> usize {
+        if self.total_size() == molecule::NUMBER_SIZE {
+            0
+        } else {
+            (molecule::unpack_number(&self.as_slice()[molecule::NUMBER_SIZE..]) as usize / 4) - 1
+        }
+    }
+    pub fn len(&self) -> usize {
+        self.item_count()
+    }
+    pub fn is_empty(&self) -> bool {
+        self.len() == 0
+    }
+    pub fn get(&self, idx: usize) -> Option<ChildScriptVec> {
+        if idx >= self.len() {
+            None
+        } else {
+            Some(self.get_unchecked(idx))
+        }
+    }
+    pub fn get_unchecked(&self, idx: usize) -> ChildScriptVec {
+        let slice = self.as_slice();
+        let start_idx = molecule::NUMBER_SIZE * (1 + idx);
+        let start = molecule::unpack_number(&slice[start_idx..]) as usize;
+        if idx == self.len() - 1 {
+            ChildScriptVec::new_unchecked(self.0.slice(start..))
+        } else {
+            let end_idx = start_idx + molecule::NUMBER_SIZE;
+            let end = molecule::unpack_number(&slice[end_idx..]) as usize;
+            ChildScriptVec::new_unchecked(self.0.slice(start..end))
+        }
+    }
+    pub fn as_reader<'r>(&'r self) -> ChildScriptVecVecReader<'r> {
+        ChildScriptVecVecReader::new_unchecked(self.as_slice())
+    }
+}
+impl molecule::prelude::Entity for ChildScriptVecVec {
+    type Builder = ChildScriptVecVecBuilder;
+    const NAME: &'static str = "ChildScriptVecVec";
+    fn new_unchecked(data: molecule::bytes::Bytes) -> Self {
+        ChildScriptVecVec(data)
+    }
+    fn as_bytes(&self) -> molecule::bytes::Bytes {
+        self.0.clone()
+    }
+    fn as_slice(&self) -> &[u8] {
+        &self.0[..]
+    }
+    fn from_slice(slice: &[u8]) -> molecule::error::VerificationResult<Self> {
+        ChildScriptVecVecReader::from_slice(slice).map(|reader| reader.to_entity())
+    }
+    fn from_compatible_slice(slice: &[u8]) -> molecule::error::VerificationResult<Self> {
+        ChildScriptVecVecReader::from_compatible_slice(slice).map(|reader| reader.to_entity())
+    }
+    fn new_builder() -> Self::Builder {
+        ::core::default::Default::default()
+    }
+    fn as_builder(self) -> Self::Builder {
+        Self::new_builder().extend(self.into_iter())
+    }
+}
+#[derive(Clone, Copy)]
+pub struct ChildScriptVecVecReader<'r>(&'r [u8]);
+impl<'r> ::core::fmt::LowerHex for ChildScriptVecVecReader<'r> {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        use molecule::hex_string;
+        if f.alternate() {
+            write!(f, "0x")?;
+        }
+        write!(f, "{}", hex_string(self.as_slice()))
+    }
+}
+impl<'r> ::core::fmt::Debug for ChildScriptVecVecReader<'r> {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        write!(f, "{}({:#x})", Self::NAME, self)
+    }
+}
+impl<'r> ::core::fmt::Display for ChildScriptVecVecReader<'r> {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        write!(f, "{} [", Self::NAME)?;
+        for i in 0..self.len() {
+            if i == 0 {
+                write!(f, "{}", self.get_unchecked(i))?;
+            } else {
+                write!(f, ", {}", self.get_unchecked(i))?;
+            }
+        }
+        write!(f, "]")
+    }
+}
+impl<'r> ChildScriptVecVecReader<'r> {
+    pub fn total_size(&self) -> usize {
+        molecule::unpack_number(self.as_slice()) as usize
+    }
+    pub fn item_count(&self) -> usize {
+        if self.total_size() == molecule::NUMBER_SIZE {
+            0
+        } else {
+            (molecule::unpack_number(&self.as_slice()[molecule::NUMBER_SIZE..]) as usize / 4) - 1
+        }
+    }
+    pub fn len(&self) -> usize {
+        self.item_count()
+    }
+    pub fn is_empty(&self) -> bool {
+        self.len() == 0
+    }
+    pub fn get(&self, idx: usize) -> Option<ChildScriptVecReader<'r>> {
+        if idx >= self.len() {
+            None
+        } else {
+            Some(self.get_unchecked(idx))
+        }
+    }
+    pub fn get_unchecked(&self, idx: usize) -> ChildScriptVecReader<'r> {
+        let slice = self.as_slice();
+        let start_idx = molecule::NUMBER_SIZE * (1 + idx);
+        let start = molecule::unpack_number(&slice[start_idx..]) as usize;
+        if idx == self.len() - 1 {
+            ChildScriptVecReader::new_unchecked(&self.as_slice()[start..])
+        } else {
+            let end_idx = start_idx + molecule::NUMBER_SIZE;
+            let end = molecule::unpack_number(&slice[end_idx..]) as usize;
+            ChildScriptVecReader::new_unchecked(&self.as_slice()[start..end])
+        }
+    }
+}
+impl<'r> molecule::prelude::Reader<'r> for ChildScriptVecVecReader<'r> {
+    type Entity = ChildScriptVecVec;
+    const NAME: &'static str = "ChildScriptVecVecReader";
+    fn to_entity(&self) -> Self::Entity {
+        Self::Entity::new_unchecked(self.as_slice().to_owned().into())
+    }
+    fn new_unchecked(slice: &'r [u8]) -> Self {
+        ChildScriptVecVecReader(slice)
+    }
+    fn as_slice(&self) -> &'r [u8] {
+        self.0
+    }
+    fn verify(slice: &[u8], compatible: bool) -> molecule::error::VerificationResult<()> {
+        use molecule::verification_error as ve;
+        let slice_len = slice.len();
+        if slice_len < molecule::NUMBER_SIZE {
+            return ve!(Self, HeaderIsBroken, molecule::NUMBER_SIZE, slice_len);
+        }
+        let total_size = molecule::unpack_number(slice) as usize;
+        if slice_len != total_size {
+            return ve!(Self, TotalSizeNotMatch, total_size, slice_len);
+        }
+        if slice_len == molecule::NUMBER_SIZE {
+            return Ok(());
+        }
+        if slice_len < molecule::NUMBER_SIZE * 2 {
+            return ve!(
+                Self,
+                TotalSizeNotMatch,
+                molecule::NUMBER_SIZE * 2,
+                slice_len
+            );
+        }
+        let offset_first = molecule::unpack_number(&slice[molecule::NUMBER_SIZE..]) as usize;
+        if offset_first % molecule::NUMBER_SIZE != 0 || offset_first < molecule::NUMBER_SIZE * 2 {
+            return ve!(Self, OffsetsNotMatch);
+        }
+        if slice_len < offset_first {
+            return ve!(Self, HeaderIsBroken, offset_first, slice_len);
+        }
+        let mut offsets: Vec<usize> = slice[molecule::NUMBER_SIZE..offset_first]
+            .chunks_exact(molecule::NUMBER_SIZE)
+            .map(|x| molecule::unpack_number(x) as usize)
+            .collect();
+        offsets.push(total_size);
+        if offsets.windows(2).any(|i| i[0] > i[1]) {
+            return ve!(Self, OffsetsNotMatch);
+        }
+        for pair in offsets.windows(2) {
+            let start = pair[0];
+            let end = pair[1];
+            ChildScriptVecReader::verify(&slice[start..end], compatible)?;
+        }
+        Ok(())
+    }
+}
+#[derive(Debug, Default)]
+pub struct ChildScriptVecVecBuilder(pub(crate) Vec<ChildScriptVec>);
+impl ChildScriptVecVecBuilder {
+    pub fn set(mut self, v: Vec<ChildScriptVec>) -> Self {
+        self.0 = v;
+        self
+    }
+    pub fn push(mut self, v: ChildScriptVec) -> Self {
+        self.0.push(v);
+        self
+    }
+    pub fn extend<T: ::core::iter::IntoIterator<Item = ChildScriptVec>>(mut self, iter: T) -> Self {
+        for elem in iter {
+            self.0.push(elem);
+        }
+        self
+    }
+    pub fn replace(&mut self, index: usize, v: ChildScriptVec) -> Option<ChildScriptVec> {
+        self.0
+            .get_mut(index)
+            .map(|item| ::core::mem::replace(item, v))
+    }
+}
+impl molecule::prelude::Builder for ChildScriptVecVecBuilder {
+    type Entity = ChildScriptVecVec;
+    const NAME: &'static str = "ChildScriptVecVecBuilder";
+    fn expected_length(&self) -> usize {
+        molecule::NUMBER_SIZE * (self.0.len() + 1)
+            + self
+                .0
+                .iter()
+                .map(|inner| inner.as_slice().len())
+                .sum::<usize>()
+    }
+    fn write<W: molecule::io::Write>(&self, writer: &mut W) -> molecule::io::Result<()> {
+        let item_count = self.0.len();
+        if item_count == 0 {
+            writer.write_all(&molecule::pack_number(
+                molecule::NUMBER_SIZE as molecule::Number,
+            ))?;
+        } else {
+            let (total_size, offsets) = self.0.iter().fold(
+                (
+                    molecule::NUMBER_SIZE * (item_count + 1),
+                    Vec::with_capacity(item_count),
+                ),
+                |(start, mut offsets), inner| {
+                    offsets.push(start);
+                    (start + inner.as_slice().len(), offsets)
+                },
+            );
+            writer.write_all(&molecule::pack_number(total_size as molecule::Number))?;
+            for offset in offsets.into_iter() {
+                writer.write_all(&molecule::pack_number(offset as molecule::Number))?;
+            }
+            for inner in self.0.iter() {
+                writer.write_all(inner.as_slice())?;
+            }
+        }
+        Ok(())
+    }
+    fn build(&self) -> Self::Entity {
+        let mut inner = Vec::with_capacity(self.expected_length());
+        self.write(&mut inner)
+            .unwrap_or_else(|_| panic!("{} build should be ok", Self::NAME));
+        ChildScriptVecVec::new_unchecked(inner.into())
+    }
+}
+pub struct ChildScriptVecVecIterator(ChildScriptVecVec, usize, usize);
+impl ::core::iter::Iterator for ChildScriptVecVecIterator {
+    type Item = ChildScriptVec;
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.1 >= self.2 {
+            None
+        } else {
+            let ret = self.0.get_unchecked(self.1);
+            self.1 += 1;
+            Some(ret)
+        }
+    }
+}
+impl ::core::iter::ExactSizeIterator for ChildScriptVecVecIterator {
+    fn len(&self) -> usize {
+        self.2 - self.1
+    }
+}
+impl ::core::iter::IntoIterator for ChildScriptVecVec {
+    type Item = ChildScriptVec;
+    type IntoIter = ChildScriptVecVecIterator;
+    fn into_iter(self) -> Self::IntoIter {
+        let len = self.len();
+        ChildScriptVecVecIterator(self, 0, len)
+    }
+}
+impl<'r> ChildScriptVecVecReader<'r> {
+    pub fn iter<'t>(&'t self) -> ChildScriptVecVecReaderIterator<'t, 'r> {
+        ChildScriptVecVecReaderIterator(&self, 0, self.len())
+    }
+}
+pub struct ChildScriptVecVecReaderIterator<'t, 'r>(&'t ChildScriptVecVecReader<'r>, usize, usize);
+impl<'t: 'r, 'r> ::core::iter::Iterator for ChildScriptVecVecReaderIterator<'t, 'r> {
+    type Item = ChildScriptVecReader<'t>;
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.1 >= self.2 {
+            None
+        } else {
+            let ret = self.0.get_unchecked(self.1);
+            self.1 += 1;
+            Some(ret)
+        }
+    }
+}
+impl<'t: 'r, 'r> ::core::iter::ExactSizeIterator for ChildScriptVecVecReaderIterator<'t, 'r> {
+    fn len(&self) -> usize {
+        self.2 - self.1
+    }
+}
+#[derive(Clone)]
+pub struct ChildScriptArray(molecule::bytes::Bytes);
+impl ::core::fmt::LowerHex for ChildScriptArray {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        use molecule::hex_string;
+        if f.alternate() {
+            write!(f, "0x")?;
+        }
+        write!(f, "{}", hex_string(self.as_slice()))
+    }
+}
+impl ::core::fmt::Debug for ChildScriptArray {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        write!(f, "{}({:#x})", Self::NAME, self)
+    }
+}
+impl ::core::fmt::Display for ChildScriptArray {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        write!(f, "{} [", Self::NAME)?;
+        for i in 0..self.len() {
+            if i == 0 {
+                write!(f, "{}", self.get_unchecked(i))?;
+            } else {
+                write!(f, ", {}", self.get_unchecked(i))?;
+            }
+        }
+        write!(f, "]")
+    }
+}
+impl ::core::default::Default for ChildScriptArray {
+    fn default() -> Self {
+        let v = molecule::bytes::Bytes::from_static(&Self::DEFAULT_VALUE);
+        ChildScriptArray::new_unchecked(v)
+    }
+}
+impl ChildScriptArray {
+    const DEFAULT_VALUE: [u8; 4] = [4, 0, 0, 0];
+    pub fn total_size(&self) -> usize {
+        molecule::unpack_number(self.as_slice()) as usize
+    }
+    pub fn item_count(&self) -> usize {
+        if self.total_size() == molecule::NUMBER_SIZE {
+            0
+        } else {
+            (molecule::unpack_number(&self.as_slice()[molecule::NUMBER_SIZE..]) as usize / 4) - 1
+        }
+    }
+    pub fn len(&self) -> usize {
+        self.item_count()
+    }
+    pub fn is_empty(&self) -> bool {
+        self.len() == 0
+    }
+    pub fn get(&self, idx: usize) -> Option<ChildScript> {
+        if idx >= self.len() {
+            None
+        } else {
+            Some(self.get_unchecked(idx))
+        }
+    }
+    pub fn get_unchecked(&self, idx: usize) -> ChildScript {
+        let slice = self.as_slice();
+        let start_idx = molecule::NUMBER_SIZE * (1 + idx);
+        let start = molecule::unpack_number(&slice[start_idx..]) as usize;
+        if idx == self.len() - 1 {
+            ChildScript::new_unchecked(self.0.slice(start..))
+        } else {
+            let end_idx = start_idx + molecule::NUMBER_SIZE;
+            let end = molecule::unpack_number(&slice[end_idx..]) as usize;
+            ChildScript::new_unchecked(self.0.slice(start..end))
+        }
+    }
+    pub fn as_reader<'r>(&'r self) -> ChildScriptArrayReader<'r> {
+        ChildScriptArrayReader::new_unchecked(self.as_slice())
+    }
+}
+impl molecule::prelude::Entity for ChildScriptArray {
+    type Builder = ChildScriptArrayBuilder;
+    const NAME: &'static str = "ChildScriptArray";
+    fn new_unchecked(data: molecule::bytes::Bytes) -> Self {
+        ChildScriptArray(data)
+    }
+    fn as_bytes(&self) -> molecule::bytes::Bytes {
+        self.0.clone()
+    }
+    fn as_slice(&self) -> &[u8] {
+        &self.0[..]
+    }
+    fn from_slice(slice: &[u8]) -> molecule::error::VerificationResult<Self> {
+        ChildScriptArrayReader::from_slice(slice).map(|reader| reader.to_entity())
+    }
+    fn from_compatible_slice(slice: &[u8]) -> molecule::error::VerificationResult<Self> {
+        ChildScriptArrayReader::from_compatible_slice(slice).map(|reader| reader.to_entity())
+    }
+    fn new_builder() -> Self::Builder {
+        ::core::default::Default::default()
+    }
+    fn as_builder(self) -> Self::Builder {
+        Self::new_builder().extend(self.into_iter())
+    }
+}
+#[derive(Clone, Copy)]
+pub struct ChildScriptArrayReader<'r>(&'r [u8]);
+impl<'r> ::core::fmt::LowerHex for ChildScriptArrayReader<'r> {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        use molecule::hex_string;
+        if f.alternate() {
+            write!(f, "0x")?;
+        }
+        write!(f, "{}", hex_string(self.as_slice()))
+    }
+}
+impl<'r> ::core::fmt::Debug for ChildScriptArrayReader<'r> {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        write!(f, "{}({:#x})", Self::NAME, self)
+    }
+}
+impl<'r> ::core::fmt::Display for ChildScriptArrayReader<'r> {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        write!(f, "{} [", Self::NAME)?;
+        for i in 0..self.len() {
+            if i == 0 {
+                write!(f, "{}", self.get_unchecked(i))?;
+            } else {
+                write!(f, ", {}", self.get_unchecked(i))?;
+            }
+        }
+        write!(f, "]")
+    }
+}
+impl<'r> ChildScriptArrayReader<'r> {
     pub fn total_size(&self) -> usize {
         molecule::unpack_number(self.as_slice()) as usize
     }
@@ -633,14 +1040,14 @@ impl<'r> ChildScriptVecReader<'r> {
         }
     }
 }
-impl<'r> molecule::prelude::Reader<'r> for ChildScriptVecReader<'r> {
-    type Entity = ChildScriptVec;
-    const NAME: &'static str = "ChildScriptVecReader";
+impl<'r> molecule::prelude::Reader<'r> for ChildScriptArrayReader<'r> {
+    type Entity = ChildScriptArray;
+    const NAME: &'static str = "ChildScriptArrayReader";
     fn to_entity(&self) -> Self::Entity {
         Self::Entity::new_unchecked(self.as_slice().to_owned().into())
     }
     fn new_unchecked(slice: &'r [u8]) -> Self {
-        ChildScriptVecReader(slice)
+        ChildScriptArrayReader(slice)
     }
     fn as_slice(&self) -> &'r [u8] {
         self.0
@@ -690,8 +1097,8 @@ impl<'r> molecule::prelude::Reader<'r> for ChildScriptVecReader<'r> {
     }
 }
 #[derive(Debug, Default)]
-pub struct ChildScriptVecBuilder(pub(crate) Vec<ChildScript>);
-impl ChildScriptVecBuilder {
+pub struct ChildScriptArrayBuilder(pub(crate) Vec<ChildScript>);
+impl ChildScriptArrayBuilder {
     pub fn set(mut self, v: Vec<ChildScript>) -> Self {
         self.0 = v;
         self
@@ -712,9 +1119,9 @@ impl ChildScriptVecBuilder {
             .map(|item| ::core::mem::replace(item, v))
     }
 }
-impl molecule::prelude::Builder for ChildScriptVecBuilder {
-    type Entity = ChildScriptVec;
-    const NAME: &'static str = "ChildScriptVecBuilder";
+impl molecule::prelude::Builder for ChildScriptArrayBuilder {
+    type Entity = ChildScriptArray;
+    const NAME: &'static str = "ChildScriptArrayBuilder";
     fn expected_length(&self) -> usize {
         molecule::NUMBER_SIZE * (self.0.len() + 1)
             + self
@@ -754,11 +1161,11 @@ impl molecule::prelude::Builder for ChildScriptVecBuilder {
         let mut inner = Vec::with_capacity(self.expected_length());
         self.write(&mut inner)
             .unwrap_or_else(|_| panic!("{} build should be ok", Self::NAME));
-        ChildScriptVec::new_unchecked(inner.into())
+        ChildScriptArray::new_unchecked(inner.into())
     }
 }
-pub struct ChildScriptVecIterator(ChildScriptVec, usize, usize);
-impl ::core::iter::Iterator for ChildScriptVecIterator {
+pub struct ChildScriptArrayIterator(ChildScriptArray, usize, usize);
+impl ::core::iter::Iterator for ChildScriptArrayIterator {
     type Item = ChildScript;
     fn next(&mut self) -> Option<Self::Item> {
         if self.1 >= self.2 {
@@ -770,26 +1177,26 @@ impl ::core::iter::Iterator for ChildScriptVecIterator {
         }
     }
 }
-impl ::core::iter::ExactSizeIterator for ChildScriptVecIterator {
+impl ::core::iter::ExactSizeIterator for ChildScriptArrayIterator {
     fn len(&self) -> usize {
         self.2 - self.1
     }
 }
-impl ::core::iter::IntoIterator for ChildScriptVec {
+impl ::core::iter::IntoIterator for ChildScriptArray {
     type Item = ChildScript;
-    type IntoIter = ChildScriptVecIterator;
+    type IntoIter = ChildScriptArrayIterator;
     fn into_iter(self) -> Self::IntoIter {
         let len = self.len();
-        ChildScriptVecIterator(self, 0, len)
+        ChildScriptArrayIterator(self, 0, len)
     }
 }
-impl<'r> ChildScriptVecReader<'r> {
-    pub fn iter<'t>(&'t self) -> ChildScriptVecReaderIterator<'t, 'r> {
-        ChildScriptVecReaderIterator(&self, 0, self.len())
+impl<'r> ChildScriptArrayReader<'r> {
+    pub fn iter<'t>(&'t self) -> ChildScriptArrayReaderIterator<'t, 'r> {
+        ChildScriptArrayReaderIterator(&self, 0, self.len())
     }
 }
-pub struct ChildScriptVecReaderIterator<'t, 'r>(&'t ChildScriptVecReader<'r>, usize, usize);
-impl<'t: 'r, 'r> ::core::iter::Iterator for ChildScriptVecReaderIterator<'t, 'r> {
+pub struct ChildScriptArrayReaderIterator<'t, 'r>(&'t ChildScriptArrayReader<'r>, usize, usize);
+impl<'t: 'r, 'r> ::core::iter::Iterator for ChildScriptArrayReaderIterator<'t, 'r> {
     type Item = ChildScriptReader<'t>;
     fn next(&mut self) -> Option<Self::Item> {
         if self.1 >= self.2 {
@@ -801,9 +1208,433 @@ impl<'t: 'r, 'r> ::core::iter::Iterator for ChildScriptVecReaderIterator<'t, 'r>
         }
     }
 }
-impl<'t: 'r, 'r> ::core::iter::ExactSizeIterator for ChildScriptVecReaderIterator<'t, 'r> {
+impl<'t: 'r, 'r> ::core::iter::ExactSizeIterator for ChildScriptArrayReaderIterator<'t, 'r> {
     fn len(&self) -> usize {
         self.2 - self.1
+    }
+}
+#[derive(Clone)]
+pub struct ChildScriptConfig(molecule::bytes::Bytes);
+impl ::core::fmt::LowerHex for ChildScriptConfig {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        use molecule::hex_string;
+        if f.alternate() {
+            write!(f, "0x")?;
+        }
+        write!(f, "{}", hex_string(self.as_slice()))
+    }
+}
+impl ::core::fmt::Debug for ChildScriptConfig {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        write!(f, "{}({:#x})", Self::NAME, self)
+    }
+}
+impl ::core::fmt::Display for ChildScriptConfig {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        write!(f, "{} {{ ", Self::NAME)?;
+        write!(f, "{}: {}", "array", self.array())?;
+        write!(f, ", {}: {}", "index", self.index())?;
+        let extra_count = self.count_extra_fields();
+        if extra_count != 0 {
+            write!(f, ", .. ({} fields)", extra_count)?;
+        }
+        write!(f, " }}")
+    }
+}
+impl ::core::default::Default for ChildScriptConfig {
+    fn default() -> Self {
+        let v = molecule::bytes::Bytes::from_static(&Self::DEFAULT_VALUE);
+        ChildScriptConfig::new_unchecked(v)
+    }
+}
+impl ChildScriptConfig {
+    const DEFAULT_VALUE: [u8; 20] = [
+        20, 0, 0, 0, 12, 0, 0, 0, 16, 0, 0, 0, 4, 0, 0, 0, 4, 0, 0, 0,
+    ];
+    pub const FIELD_COUNT: usize = 2;
+    pub fn total_size(&self) -> usize {
+        molecule::unpack_number(self.as_slice()) as usize
+    }
+    pub fn field_count(&self) -> usize {
+        if self.total_size() == molecule::NUMBER_SIZE {
+            0
+        } else {
+            (molecule::unpack_number(&self.as_slice()[molecule::NUMBER_SIZE..]) as usize / 4) - 1
+        }
+    }
+    pub fn count_extra_fields(&self) -> usize {
+        self.field_count() - Self::FIELD_COUNT
+    }
+    pub fn has_extra_fields(&self) -> bool {
+        Self::FIELD_COUNT != self.field_count()
+    }
+    pub fn array(&self) -> ChildScriptArray {
+        let slice = self.as_slice();
+        let start = molecule::unpack_number(&slice[4..]) as usize;
+        let end = molecule::unpack_number(&slice[8..]) as usize;
+        ChildScriptArray::new_unchecked(self.0.slice(start..end))
+    }
+    pub fn index(&self) -> ChildScriptVecVec {
+        let slice = self.as_slice();
+        let start = molecule::unpack_number(&slice[8..]) as usize;
+        if self.has_extra_fields() {
+            let end = molecule::unpack_number(&slice[12..]) as usize;
+            ChildScriptVecVec::new_unchecked(self.0.slice(start..end))
+        } else {
+            ChildScriptVecVec::new_unchecked(self.0.slice(start..))
+        }
+    }
+    pub fn as_reader<'r>(&'r self) -> ChildScriptConfigReader<'r> {
+        ChildScriptConfigReader::new_unchecked(self.as_slice())
+    }
+}
+impl molecule::prelude::Entity for ChildScriptConfig {
+    type Builder = ChildScriptConfigBuilder;
+    const NAME: &'static str = "ChildScriptConfig";
+    fn new_unchecked(data: molecule::bytes::Bytes) -> Self {
+        ChildScriptConfig(data)
+    }
+    fn as_bytes(&self) -> molecule::bytes::Bytes {
+        self.0.clone()
+    }
+    fn as_slice(&self) -> &[u8] {
+        &self.0[..]
+    }
+    fn from_slice(slice: &[u8]) -> molecule::error::VerificationResult<Self> {
+        ChildScriptConfigReader::from_slice(slice).map(|reader| reader.to_entity())
+    }
+    fn from_compatible_slice(slice: &[u8]) -> molecule::error::VerificationResult<Self> {
+        ChildScriptConfigReader::from_compatible_slice(slice).map(|reader| reader.to_entity())
+    }
+    fn new_builder() -> Self::Builder {
+        ::core::default::Default::default()
+    }
+    fn as_builder(self) -> Self::Builder {
+        Self::new_builder().array(self.array()).index(self.index())
+    }
+}
+#[derive(Clone, Copy)]
+pub struct ChildScriptConfigReader<'r>(&'r [u8]);
+impl<'r> ::core::fmt::LowerHex for ChildScriptConfigReader<'r> {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        use molecule::hex_string;
+        if f.alternate() {
+            write!(f, "0x")?;
+        }
+        write!(f, "{}", hex_string(self.as_slice()))
+    }
+}
+impl<'r> ::core::fmt::Debug for ChildScriptConfigReader<'r> {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        write!(f, "{}({:#x})", Self::NAME, self)
+    }
+}
+impl<'r> ::core::fmt::Display for ChildScriptConfigReader<'r> {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        write!(f, "{} {{ ", Self::NAME)?;
+        write!(f, "{}: {}", "array", self.array())?;
+        write!(f, ", {}: {}", "index", self.index())?;
+        let extra_count = self.count_extra_fields();
+        if extra_count != 0 {
+            write!(f, ", .. ({} fields)", extra_count)?;
+        }
+        write!(f, " }}")
+    }
+}
+impl<'r> ChildScriptConfigReader<'r> {
+    pub const FIELD_COUNT: usize = 2;
+    pub fn total_size(&self) -> usize {
+        molecule::unpack_number(self.as_slice()) as usize
+    }
+    pub fn field_count(&self) -> usize {
+        if self.total_size() == molecule::NUMBER_SIZE {
+            0
+        } else {
+            (molecule::unpack_number(&self.as_slice()[molecule::NUMBER_SIZE..]) as usize / 4) - 1
+        }
+    }
+    pub fn count_extra_fields(&self) -> usize {
+        self.field_count() - Self::FIELD_COUNT
+    }
+    pub fn has_extra_fields(&self) -> bool {
+        Self::FIELD_COUNT != self.field_count()
+    }
+    pub fn array(&self) -> ChildScriptArrayReader<'r> {
+        let slice = self.as_slice();
+        let start = molecule::unpack_number(&slice[4..]) as usize;
+        let end = molecule::unpack_number(&slice[8..]) as usize;
+        ChildScriptArrayReader::new_unchecked(&self.as_slice()[start..end])
+    }
+    pub fn index(&self) -> ChildScriptVecVecReader<'r> {
+        let slice = self.as_slice();
+        let start = molecule::unpack_number(&slice[8..]) as usize;
+        if self.has_extra_fields() {
+            let end = molecule::unpack_number(&slice[12..]) as usize;
+            ChildScriptVecVecReader::new_unchecked(&self.as_slice()[start..end])
+        } else {
+            ChildScriptVecVecReader::new_unchecked(&self.as_slice()[start..])
+        }
+    }
+}
+impl<'r> molecule::prelude::Reader<'r> for ChildScriptConfigReader<'r> {
+    type Entity = ChildScriptConfig;
+    const NAME: &'static str = "ChildScriptConfigReader";
+    fn to_entity(&self) -> Self::Entity {
+        Self::Entity::new_unchecked(self.as_slice().to_owned().into())
+    }
+    fn new_unchecked(slice: &'r [u8]) -> Self {
+        ChildScriptConfigReader(slice)
+    }
+    fn as_slice(&self) -> &'r [u8] {
+        self.0
+    }
+    fn verify(slice: &[u8], compatible: bool) -> molecule::error::VerificationResult<()> {
+        use molecule::verification_error as ve;
+        let slice_len = slice.len();
+        if slice_len < molecule::NUMBER_SIZE {
+            return ve!(Self, HeaderIsBroken, molecule::NUMBER_SIZE, slice_len);
+        }
+        let total_size = molecule::unpack_number(slice) as usize;
+        if slice_len != total_size {
+            return ve!(Self, TotalSizeNotMatch, total_size, slice_len);
+        }
+        if slice_len < molecule::NUMBER_SIZE * 2 {
+            return ve!(Self, HeaderIsBroken, molecule::NUMBER_SIZE * 2, slice_len);
+        }
+        let offset_first = molecule::unpack_number(&slice[molecule::NUMBER_SIZE..]) as usize;
+        if offset_first % molecule::NUMBER_SIZE != 0 || offset_first < molecule::NUMBER_SIZE * 2 {
+            return ve!(Self, OffsetsNotMatch);
+        }
+        if slice_len < offset_first {
+            return ve!(Self, HeaderIsBroken, offset_first, slice_len);
+        }
+        let field_count = offset_first / molecule::NUMBER_SIZE - 1;
+        if field_count < Self::FIELD_COUNT {
+            return ve!(Self, FieldCountNotMatch, Self::FIELD_COUNT, field_count);
+        } else if !compatible && field_count > Self::FIELD_COUNT {
+            return ve!(Self, FieldCountNotMatch, Self::FIELD_COUNT, field_count);
+        };
+        let mut offsets: Vec<usize> = slice[molecule::NUMBER_SIZE..offset_first]
+            .chunks_exact(molecule::NUMBER_SIZE)
+            .map(|x| molecule::unpack_number(x) as usize)
+            .collect();
+        offsets.push(total_size);
+        if offsets.windows(2).any(|i| i[0] > i[1]) {
+            return ve!(Self, OffsetsNotMatch);
+        }
+        ChildScriptArrayReader::verify(&slice[offsets[0]..offsets[1]], compatible)?;
+        ChildScriptVecVecReader::verify(&slice[offsets[1]..offsets[2]], compatible)?;
+        Ok(())
+    }
+}
+#[derive(Debug, Default)]
+pub struct ChildScriptConfigBuilder {
+    pub(crate) array: ChildScriptArray,
+    pub(crate) index: ChildScriptVecVec,
+}
+impl ChildScriptConfigBuilder {
+    pub const FIELD_COUNT: usize = 2;
+    pub fn array(mut self, v: ChildScriptArray) -> Self {
+        self.array = v;
+        self
+    }
+    pub fn index(mut self, v: ChildScriptVecVec) -> Self {
+        self.index = v;
+        self
+    }
+}
+impl molecule::prelude::Builder for ChildScriptConfigBuilder {
+    type Entity = ChildScriptConfig;
+    const NAME: &'static str = "ChildScriptConfigBuilder";
+    fn expected_length(&self) -> usize {
+        molecule::NUMBER_SIZE * (Self::FIELD_COUNT + 1)
+            + self.array.as_slice().len()
+            + self.index.as_slice().len()
+    }
+    fn write<W: molecule::io::Write>(&self, writer: &mut W) -> molecule::io::Result<()> {
+        let mut total_size = molecule::NUMBER_SIZE * (Self::FIELD_COUNT + 1);
+        let mut offsets = Vec::with_capacity(Self::FIELD_COUNT);
+        offsets.push(total_size);
+        total_size += self.array.as_slice().len();
+        offsets.push(total_size);
+        total_size += self.index.as_slice().len();
+        writer.write_all(&molecule::pack_number(total_size as molecule::Number))?;
+        for offset in offsets.into_iter() {
+            writer.write_all(&molecule::pack_number(offset as molecule::Number))?;
+        }
+        writer.write_all(self.array.as_slice())?;
+        writer.write_all(self.index.as_slice())?;
+        Ok(())
+    }
+    fn build(&self) -> Self::Entity {
+        let mut inner = Vec::with_capacity(self.expected_length());
+        self.write(&mut inner)
+            .unwrap_or_else(|_| panic!("{} build should be ok", Self::NAME));
+        ChildScriptConfig::new_unchecked(inner.into())
+    }
+}
+#[derive(Clone)]
+pub struct ChildScriptConfigOpt(molecule::bytes::Bytes);
+impl ::core::fmt::LowerHex for ChildScriptConfigOpt {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        use molecule::hex_string;
+        if f.alternate() {
+            write!(f, "0x")?;
+        }
+        write!(f, "{}", hex_string(self.as_slice()))
+    }
+}
+impl ::core::fmt::Debug for ChildScriptConfigOpt {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        write!(f, "{}({:#x})", Self::NAME, self)
+    }
+}
+impl ::core::fmt::Display for ChildScriptConfigOpt {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        if let Some(v) = self.to_opt() {
+            write!(f, "{}(Some({}))", Self::NAME, v)
+        } else {
+            write!(f, "{}(None)", Self::NAME)
+        }
+    }
+}
+impl ::core::default::Default for ChildScriptConfigOpt {
+    fn default() -> Self {
+        let v = molecule::bytes::Bytes::from_static(&Self::DEFAULT_VALUE);
+        ChildScriptConfigOpt::new_unchecked(v)
+    }
+}
+impl ChildScriptConfigOpt {
+    const DEFAULT_VALUE: [u8; 0] = [];
+    pub fn is_none(&self) -> bool {
+        self.0.is_empty()
+    }
+    pub fn is_some(&self) -> bool {
+        !self.0.is_empty()
+    }
+    pub fn to_opt(&self) -> Option<ChildScriptConfig> {
+        if self.is_none() {
+            None
+        } else {
+            Some(ChildScriptConfig::new_unchecked(self.0.clone()))
+        }
+    }
+    pub fn as_reader<'r>(&'r self) -> ChildScriptConfigOptReader<'r> {
+        ChildScriptConfigOptReader::new_unchecked(self.as_slice())
+    }
+}
+impl molecule::prelude::Entity for ChildScriptConfigOpt {
+    type Builder = ChildScriptConfigOptBuilder;
+    const NAME: &'static str = "ChildScriptConfigOpt";
+    fn new_unchecked(data: molecule::bytes::Bytes) -> Self {
+        ChildScriptConfigOpt(data)
+    }
+    fn as_bytes(&self) -> molecule::bytes::Bytes {
+        self.0.clone()
+    }
+    fn as_slice(&self) -> &[u8] {
+        &self.0[..]
+    }
+    fn from_slice(slice: &[u8]) -> molecule::error::VerificationResult<Self> {
+        ChildScriptConfigOptReader::from_slice(slice).map(|reader| reader.to_entity())
+    }
+    fn from_compatible_slice(slice: &[u8]) -> molecule::error::VerificationResult<Self> {
+        ChildScriptConfigOptReader::from_compatible_slice(slice).map(|reader| reader.to_entity())
+    }
+    fn new_builder() -> Self::Builder {
+        ::core::default::Default::default()
+    }
+    fn as_builder(self) -> Self::Builder {
+        Self::new_builder().set(self.to_opt())
+    }
+}
+#[derive(Clone, Copy)]
+pub struct ChildScriptConfigOptReader<'r>(&'r [u8]);
+impl<'r> ::core::fmt::LowerHex for ChildScriptConfigOptReader<'r> {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        use molecule::hex_string;
+        if f.alternate() {
+            write!(f, "0x")?;
+        }
+        write!(f, "{}", hex_string(self.as_slice()))
+    }
+}
+impl<'r> ::core::fmt::Debug for ChildScriptConfigOptReader<'r> {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        write!(f, "{}({:#x})", Self::NAME, self)
+    }
+}
+impl<'r> ::core::fmt::Display for ChildScriptConfigOptReader<'r> {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+        if let Some(v) = self.to_opt() {
+            write!(f, "{}(Some({}))", Self::NAME, v)
+        } else {
+            write!(f, "{}(None)", Self::NAME)
+        }
+    }
+}
+impl<'r> ChildScriptConfigOptReader<'r> {
+    pub fn is_none(&self) -> bool {
+        self.0.is_empty()
+    }
+    pub fn is_some(&self) -> bool {
+        !self.0.is_empty()
+    }
+    pub fn to_opt(&self) -> Option<ChildScriptConfigReader<'r>> {
+        if self.is_none() {
+            None
+        } else {
+            Some(ChildScriptConfigReader::new_unchecked(self.as_slice()))
+        }
+    }
+}
+impl<'r> molecule::prelude::Reader<'r> for ChildScriptConfigOptReader<'r> {
+    type Entity = ChildScriptConfigOpt;
+    const NAME: &'static str = "ChildScriptConfigOptReader";
+    fn to_entity(&self) -> Self::Entity {
+        Self::Entity::new_unchecked(self.as_slice().to_owned().into())
+    }
+    fn new_unchecked(slice: &'r [u8]) -> Self {
+        ChildScriptConfigOptReader(slice)
+    }
+    fn as_slice(&self) -> &'r [u8] {
+        self.0
+    }
+    fn verify(slice: &[u8], compatible: bool) -> molecule::error::VerificationResult<()> {
+        if !slice.is_empty() {
+            ChildScriptConfigReader::verify(&slice[..], compatible)?;
+        }
+        Ok(())
+    }
+}
+#[derive(Debug, Default)]
+pub struct ChildScriptConfigOptBuilder(pub(crate) Option<ChildScriptConfig>);
+impl ChildScriptConfigOptBuilder {
+    pub fn set(mut self, v: Option<ChildScriptConfig>) -> Self {
+        self.0 = v;
+        self
+    }
+}
+impl molecule::prelude::Builder for ChildScriptConfigOptBuilder {
+    type Entity = ChildScriptConfigOpt;
+    const NAME: &'static str = "ChildScriptConfigOptBuilder";
+    fn expected_length(&self) -> usize {
+        self.0
+            .as_ref()
+            .map(|ref inner| inner.as_slice().len())
+            .unwrap_or(0)
+    }
+    fn write<W: molecule::io::Write>(&self, writer: &mut W) -> molecule::io::Result<()> {
+        self.0
+            .as_ref()
+            .map(|ref inner| writer.write_all(inner.as_slice()))
+            .unwrap_or(Ok(()))
+    }
+    fn build(&self) -> Self::Entity {
+        let mut inner = Vec::with_capacity(self.expected_length());
+        self.write(&mut inner)
+            .unwrap_or_else(|_| panic!("{} build should be ok", Self::NAME));
+        ChildScriptConfigOpt::new_unchecked(inner.into())
     }
 }
 #[derive(Clone)]
@@ -825,9 +1656,9 @@ impl ::core::fmt::Debug for CombineLockWitness {
 impl ::core::fmt::Display for CombineLockWitness {
     fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
         write!(f, "{} {{ ", Self::NAME)?;
-        write!(f, "{}: {}", "witness_base_index", self.witness_base_index())?;
-        write!(f, ", {}: {}", "proof", self.proof())?;
-        write!(f, ", {}: {}", "scripts", self.scripts())?;
+        write!(f, "{}: {}", "index", self.index())?;
+        write!(f, ", {}: {}", "inner_witness", self.inner_witness())?;
+        write!(f, ", {}: {}", "script_config", self.script_config())?;
         let extra_count = self.count_extra_fields();
         if extra_count != 0 {
             write!(f, ", .. ({} fields)", extra_count)?;
@@ -837,13 +1668,14 @@ impl ::core::fmt::Display for CombineLockWitness {
 }
 impl ::core::default::Default for CombineLockWitness {
     fn default() -> Self {
-        let v: Vec<u8> = vec![
-            26, 0, 0, 0, 16, 0, 0, 0, 18, 0, 0, 0, 22, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 0, 0, 0,
-        ];
-        CombineLockWitness::new_unchecked(v.into())
+        let v = molecule::bytes::Bytes::from_static(&Self::DEFAULT_VALUE);
+        CombineLockWitness::new_unchecked(v)
     }
 }
 impl CombineLockWitness {
+    const DEFAULT_VALUE: [u8; 21] = [
+        21, 0, 0, 0, 16, 0, 0, 0, 17, 0, 0, 0, 21, 0, 0, 0, 0, 4, 0, 0, 0,
+    ];
     pub const FIELD_COUNT: usize = 3;
     pub fn total_size(&self) -> usize {
         molecule::unpack_number(self.as_slice()) as usize
@@ -861,26 +1693,26 @@ impl CombineLockWitness {
     pub fn has_extra_fields(&self) -> bool {
         Self::FIELD_COUNT != self.field_count()
     }
-    pub fn witness_base_index(&self) -> Uint16 {
+    pub fn index(&self) -> Byte {
         let slice = self.as_slice();
         let start = molecule::unpack_number(&slice[4..]) as usize;
         let end = molecule::unpack_number(&slice[8..]) as usize;
-        Uint16::new_unchecked(self.0.slice(start..end))
+        Byte::new_unchecked(self.0.slice(start..end))
     }
-    pub fn proof(&self) -> Bytes {
+    pub fn inner_witness(&self) -> BytesVec {
         let slice = self.as_slice();
         let start = molecule::unpack_number(&slice[8..]) as usize;
         let end = molecule::unpack_number(&slice[12..]) as usize;
-        Bytes::new_unchecked(self.0.slice(start..end))
+        BytesVec::new_unchecked(self.0.slice(start..end))
     }
-    pub fn scripts(&self) -> ChildScriptVec {
+    pub fn script_config(&self) -> ChildScriptConfigOpt {
         let slice = self.as_slice();
         let start = molecule::unpack_number(&slice[12..]) as usize;
         if self.has_extra_fields() {
             let end = molecule::unpack_number(&slice[16..]) as usize;
-            ChildScriptVec::new_unchecked(self.0.slice(start..end))
+            ChildScriptConfigOpt::new_unchecked(self.0.slice(start..end))
         } else {
-            ChildScriptVec::new_unchecked(self.0.slice(start..))
+            ChildScriptConfigOpt::new_unchecked(self.0.slice(start..))
         }
     }
     pub fn as_reader<'r>(&'r self) -> CombineLockWitnessReader<'r> {
@@ -910,9 +1742,9 @@ impl molecule::prelude::Entity for CombineLockWitness {
     }
     fn as_builder(self) -> Self::Builder {
         Self::new_builder()
-            .witness_base_index(self.witness_base_index())
-            .proof(self.proof())
-            .scripts(self.scripts())
+            .index(self.index())
+            .inner_witness(self.inner_witness())
+            .script_config(self.script_config())
     }
 }
 #[derive(Clone, Copy)]
@@ -934,9 +1766,9 @@ impl<'r> ::core::fmt::Debug for CombineLockWitnessReader<'r> {
 impl<'r> ::core::fmt::Display for CombineLockWitnessReader<'r> {
     fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
         write!(f, "{} {{ ", Self::NAME)?;
-        write!(f, "{}: {}", "witness_base_index", self.witness_base_index())?;
-        write!(f, ", {}: {}", "proof", self.proof())?;
-        write!(f, ", {}: {}", "scripts", self.scripts())?;
+        write!(f, "{}: {}", "index", self.index())?;
+        write!(f, ", {}: {}", "inner_witness", self.inner_witness())?;
+        write!(f, ", {}: {}", "script_config", self.script_config())?;
         let extra_count = self.count_extra_fields();
         if extra_count != 0 {
             write!(f, ", .. ({} fields)", extra_count)?;
@@ -962,26 +1794,26 @@ impl<'r> CombineLockWitnessReader<'r> {
     pub fn has_extra_fields(&self) -> bool {
         Self::FIELD_COUNT != self.field_count()
     }
-    pub fn witness_base_index(&self) -> Uint16Reader<'r> {
+    pub fn index(&self) -> ByteReader<'r> {
         let slice = self.as_slice();
         let start = molecule::unpack_number(&slice[4..]) as usize;
         let end = molecule::unpack_number(&slice[8..]) as usize;
-        Uint16Reader::new_unchecked(&self.as_slice()[start..end])
+        ByteReader::new_unchecked(&self.as_slice()[start..end])
     }
-    pub fn proof(&self) -> BytesReader<'r> {
+    pub fn inner_witness(&self) -> BytesVecReader<'r> {
         let slice = self.as_slice();
         let start = molecule::unpack_number(&slice[8..]) as usize;
         let end = molecule::unpack_number(&slice[12..]) as usize;
-        BytesReader::new_unchecked(&self.as_slice()[start..end])
+        BytesVecReader::new_unchecked(&self.as_slice()[start..end])
     }
-    pub fn scripts(&self) -> ChildScriptVecReader<'r> {
+    pub fn script_config(&self) -> ChildScriptConfigOptReader<'r> {
         let slice = self.as_slice();
         let start = molecule::unpack_number(&slice[12..]) as usize;
         if self.has_extra_fields() {
             let end = molecule::unpack_number(&slice[16..]) as usize;
-            ChildScriptVecReader::new_unchecked(&self.as_slice()[start..end])
+            ChildScriptConfigOptReader::new_unchecked(&self.as_slice()[start..end])
         } else {
-            ChildScriptVecReader::new_unchecked(&self.as_slice()[start..])
+            ChildScriptConfigOptReader::new_unchecked(&self.as_slice()[start..])
         }
     }
 }
@@ -1007,9 +1839,6 @@ impl<'r> molecule::prelude::Reader<'r> for CombineLockWitnessReader<'r> {
         if slice_len != total_size {
             return ve!(Self, TotalSizeNotMatch, total_size, slice_len);
         }
-        if slice_len == molecule::NUMBER_SIZE && Self::FIELD_COUNT == 0 {
-            return Ok(());
-        }
         if slice_len < molecule::NUMBER_SIZE * 2 {
             return ve!(Self, HeaderIsBroken, molecule::NUMBER_SIZE * 2, slice_len);
         }
@@ -1034,30 +1863,30 @@ impl<'r> molecule::prelude::Reader<'r> for CombineLockWitnessReader<'r> {
         if offsets.windows(2).any(|i| i[0] > i[1]) {
             return ve!(Self, OffsetsNotMatch);
         }
-        Uint16Reader::verify(&slice[offsets[0]..offsets[1]], compatible)?;
-        BytesReader::verify(&slice[offsets[1]..offsets[2]], compatible)?;
-        ChildScriptVecReader::verify(&slice[offsets[2]..offsets[3]], compatible)?;
+        ByteReader::verify(&slice[offsets[0]..offsets[1]], compatible)?;
+        BytesVecReader::verify(&slice[offsets[1]..offsets[2]], compatible)?;
+        ChildScriptConfigOptReader::verify(&slice[offsets[2]..offsets[3]], compatible)?;
         Ok(())
     }
 }
 #[derive(Debug, Default)]
 pub struct CombineLockWitnessBuilder {
-    pub(crate) witness_base_index: Uint16,
-    pub(crate) proof: Bytes,
-    pub(crate) scripts: ChildScriptVec,
+    pub(crate) index: Byte,
+    pub(crate) inner_witness: BytesVec,
+    pub(crate) script_config: ChildScriptConfigOpt,
 }
 impl CombineLockWitnessBuilder {
     pub const FIELD_COUNT: usize = 3;
-    pub fn witness_base_index(mut self, v: Uint16) -> Self {
-        self.witness_base_index = v;
+    pub fn index(mut self, v: Byte) -> Self {
+        self.index = v;
         self
     }
-    pub fn proof(mut self, v: Bytes) -> Self {
-        self.proof = v;
+    pub fn inner_witness(mut self, v: BytesVec) -> Self {
+        self.inner_witness = v;
         self
     }
-    pub fn scripts(mut self, v: ChildScriptVec) -> Self {
-        self.scripts = v;
+    pub fn script_config(mut self, v: ChildScriptConfigOpt) -> Self {
+        self.script_config = v;
         self
     }
 }
@@ -1066,26 +1895,26 @@ impl molecule::prelude::Builder for CombineLockWitnessBuilder {
     const NAME: &'static str = "CombineLockWitnessBuilder";
     fn expected_length(&self) -> usize {
         molecule::NUMBER_SIZE * (Self::FIELD_COUNT + 1)
-            + self.witness_base_index.as_slice().len()
-            + self.proof.as_slice().len()
-            + self.scripts.as_slice().len()
+            + self.index.as_slice().len()
+            + self.inner_witness.as_slice().len()
+            + self.script_config.as_slice().len()
     }
     fn write<W: molecule::io::Write>(&self, writer: &mut W) -> molecule::io::Result<()> {
         let mut total_size = molecule::NUMBER_SIZE * (Self::FIELD_COUNT + 1);
         let mut offsets = Vec::with_capacity(Self::FIELD_COUNT);
         offsets.push(total_size);
-        total_size += self.witness_base_index.as_slice().len();
+        total_size += self.index.as_slice().len();
         offsets.push(total_size);
-        total_size += self.proof.as_slice().len();
+        total_size += self.inner_witness.as_slice().len();
         offsets.push(total_size);
-        total_size += self.scripts.as_slice().len();
+        total_size += self.script_config.as_slice().len();
         writer.write_all(&molecule::pack_number(total_size as molecule::Number))?;
         for offset in offsets.into_iter() {
             writer.write_all(&molecule::pack_number(offset as molecule::Number))?;
         }
-        writer.write_all(self.witness_base_index.as_slice())?;
-        writer.write_all(self.proof.as_slice())?;
-        writer.write_all(self.scripts.as_slice())?;
+        writer.write_all(self.index.as_slice())?;
+        writer.write_all(self.inner_witness.as_slice())?;
+        writer.write_all(self.script_config.as_slice())?;
         Ok(())
     }
     fn build(&self) -> Self::Entity {
