+++ ckb-debugger-tests/templates/cl-child-script.json
@@ -25,7 +25,7 @@
           },
           "type": "{{ def_type combine_lock }}"
         },
-        "data": "0x{{ data ../../target/riscv64imac-unknown-none-elf/release/ckb-combine-lock-v2 }}"
+        "data": "0x{{ data ../../target/riscv64imac-unknown-none-elf/release/ckb-combine-lock }}"
       },
       {
         "output": {
