--- contracts/ckb-combine-lock-v2/rustfmt.toml
@@ -1,2 +1,0 @@
-max_width = 120
-chain_width = 120
