--- contracts/ckb-combine-lock/.cargo/config
@@ -1,2 +1,0 @@
-[build]
-target = "riscv64imac-unknown-none-elf"
