--- contracts/ckb-combine-lock/src/constant.rs
@@ -1,5 +1,0 @@
-pub const ARGS_SIZE: usize = 33;
-pub const SMT_VALUE: [u8; 32] = [
-    1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-];
-pub const SMT_SIZE: usize = 32;
