+++ contracts/ckb-combine-lock/src/error.rs
@@ -1,21 +1,16 @@
 use ckb_std::error::SysError;
-
 /// Error
 #[repr(i8)]
-#[derive(Debug)]
 pub enum Error {
     IndexOutOfBound = 1,
     ItemMissing,
     LengthNotEnough,
     Encoding,
-    // combine lock errors starts from 80
-    WrongArgs = 80,
-    WrongInfoCell,
-    WrongWitnessFormat,
-    SmtVerifyFailed,
-    WrongMolecule,
-    WrongHex,
-    ExecError,
+    // Add customized errors here...
+    WrongFormat = 80,
+    WrongScriptConfigHash,
+    WrongHashType,
+    UnlockFailed,
 }
 
 impl From<SysError> for Error {
@@ -28,5 +23,17 @@ impl From<SysError> for Error {
             Encoding => Self::Encoding,
             Unknown(err_code) => panic!("unexpected sys error {}", err_code),
         }
+    }
+}
+
+impl From<molecule::error::VerificationError> for Error {
+    fn from(_: molecule::error::VerificationError) -> Self {
+        Self::WrongFormat
+    }
+}
+
+impl From<hex::FromHexError> for Error {
+    fn from(_: hex::FromHexError) -> Self {
+        Self::WrongFormat
     }
 }
