+++ contracts/ckb-combine-lock/src/main.rs
@@ -7,6 +7,12 @@
 #![no_std]
 #![cfg_attr(not(test), no_main)]
 
+// define modules
+mod blake2b;
+mod entry;
+mod error;
+use ckb_combine_lock_common::logger;
+
 #[cfg(test)]
 extern crate alloc;
 
@@ -17,26 +23,12 @@ ckb_std::entry!(program_entry);
 #[cfg(not(test))]
 default_alloc!();
 
-// define modules
-mod blake2b;
-mod constant;
-mod entry;
-mod error;
-
-use ckb_combine_lock_common::logger;
-use log::warn;
-
 /// program entry
-///
-///  Both `argc` and `argv` can be omitted.
-fn program_entry() -> i8 {
+pub fn program_entry() -> i8 {
     drop(logger::init());
     // Call main function and return error code
     match entry::main() {
         Ok(_) => 0,
-        Err(err) => {
-            warn!("script exit with error: {:?}", err);
-            err as i8
-        }
+        Err(err) => err as i8,
     }
 }
